[cli] Tweak file layout
