+++ .gitattributes
@@ -1,5 +1,8 @@
 # Auto detect text files and perform LF normalization
 * text=auto
-.yarn/** linguist-generated=true
-samlang-core-parser/** linguist-generated=true
+.yarn/plugins/** linguist-generated=true
+.yarn/releases/** linguist-generated=true
+.yarn/sdks/** linguist-generated=true
+configurations/* linguist-generated=true
+samlang-core-parser-generated/** linguist-generated=true
 samlang-cli/samlang-runtime/* linguist-generated=true
